refactoring packages structure
