--- app/src/main/java/com/ringov/yamblzweather/db/Database.java
@@ -1,26 +1,0 @@
-package com.ringov.yamblzweather.db;
-
-import com.ringov.yamblzweather.db.model.DBWeather;
-
-/**
- * Created by ringov on 15.07.17.
- */
-
-public class Database {
-    private static Database instance;
-
-    public static Database getInstance() {
-        if (instance == null) {
-            instance = new Database();
-        }
-        return instance;
-    }
-
-    public Database() {
-
-    }
-
-    public DBWeather getWeather(int cityId) {
-        return new DBWeather();
-    }
-}
