--- app/src/main/java/com/ringov/yamblzweather/db/model/DBWeather.java
@@ -1,8 +1,0 @@
-package com.ringov.yamblzweather.db.model;
-
-/**
- * Created by ringov on 15.07.17.
- */
-
-public class DBWeather {
-}
