--- app/src/main/java/com/ringov/yamblzweather/internet/content/Clouds.java
@@ -1,21 +1,0 @@
-
-package com.ringov.yamblzweather.internet.content;
-
-import com.google.gson.annotations.Expose;
-import com.google.gson.annotations.SerializedName;
-
-public class Clouds {
-
-    @SerializedName("all")
-    @Expose
-    private Integer all;
-
-    public Integer getAll() {
-        return all;
-    }
-
-    public void setAll(Integer all) {
-        this.all = all;
-    }
-
-}
