--- app/src/main/java/com/ringov/yamblzweather/internet/content/Coord.java
@@ -1,32 +1,0 @@
-
-package com.ringov.yamblzweather.internet.content;
-
-import com.google.gson.annotations.Expose;
-import com.google.gson.annotations.SerializedName;
-
-public class Coord {
-
-    @SerializedName("lon")
-    @Expose
-    private Double lon;
-    @SerializedName("lat")
-    @Expose
-    private Double lat;
-
-    public Double getLon() {
-        return lon;
-    }
-
-    public void setLon(Double lon) {
-        this.lon = lon;
-    }
-
-    public Double getLat() {
-        return lat;
-    }
-
-    public void setLat(Double lat) {
-        this.lat = lat;
-    }
-
-}
