--- app/src/main/java/com/ringov/yamblzweather/internet/content/Wind.java
@@ -1,32 +1,0 @@
-
-package com.ringov.yamblzweather.internet.content;
-
-import com.google.gson.annotations.Expose;
-import com.google.gson.annotations.SerializedName;
-
-public class Wind {
-
-    @SerializedName("speed")
-    @Expose
-    private Integer speed;
-    @SerializedName("deg")
-    @Expose
-    private Integer deg;
-
-    public Integer getSpeed() {
-        return speed;
-    }
-
-    public void setSpeed(Integer speed) {
-        this.speed = speed;
-    }
-
-    public Integer getDeg() {
-        return deg;
-    }
-
-    public void setDeg(Integer deg) {
-        this.deg = deg;
-    }
-
-}
