--- app/src/main/java/com/ringov/yamblzweather/model/about/AboutRepository.java
@@ -1,10 +1,0 @@
-package com.ringov.yamblzweather.model.about;
-
-import com.ringov.yamblzweather.model.base.BaseRepository;
-
-/**
- * Created by ringov on 12.07.17.
- */
-
-public interface AboutRepository extends BaseRepository {
-}
