--- app/src/main/java/com/ringov/yamblzweather/model/about/AboutRepositoryImpl.java
@@ -1,13 +1,0 @@
-package com.ringov.yamblzweather.model.about;
-
-import com.ringov.yamblzweather.model.base.BaseRepositoryImpl;
-
-import javax.inject.Singleton;
-
-/**
- * Created by ringov on 12.07.17.
- */
-
-@Singleton
-public class AboutRepositoryImpl extends BaseRepositoryImpl implements AboutRepository {
-}
