--- app/src/main/java/com/ringov/yamblzweather/model/base/BaseRepository.java
@@ -1,8 +1,0 @@
-package com.ringov.yamblzweather.model.base;
-
-/**
- * Created by ringov on 12.07.17.
- */
-
-public interface BaseRepository {
-}
