--- app/src/main/java/com/ringov/yamblzweather/model/base/BaseRepositoryImpl.java
@@ -1,8 +1,0 @@
-package com.ringov.yamblzweather.model.base;
-
-/**
- * Created by ringov on 12.07.17.
- */
-
-public class BaseRepositoryImpl implements BaseRepository {
-}
