+++ app/src/main/java/com/ringov/yamblzweather/model/db/model/DBWeather.java
@@ -1,0 +1,8 @@
+package com.ringov.yamblzweather.model.db.model;
+
+/**
+ * Created by ringov on 15.07.17.
+ */
+
+public class DBWeather {
+}
