+++ app/src/main/java/com/ringov/yamblzweather/model/internet/content/Wind.java
@@ -1,0 +1,32 @@
+
+package com.ringov.yamblzweather.model.internet.content;
+
+import com.google.gson.annotations.Expose;
+import com.google.gson.annotations.SerializedName;
+
+public class Wind {
+
+    @SerializedName("speed")
+    @Expose
+    private Integer speed;
+    @SerializedName("deg")
+    @Expose
+    private Integer deg;
+
+    public Integer getSpeed() {
+        return speed;
+    }
+
+    public void setSpeed(Integer speed) {
+        this.speed = speed;
+    }
+
+    public Integer getDeg() {
+        return deg;
+    }
+
+    public void setDeg(Integer deg) {
+        this.deg = deg;
+    }
+
+}
