+++ app/src/main/java/com/ringov/yamblzweather/model/repositories/about/AboutRepository.java
@@ -1,0 +1,10 @@
+package com.ringov.yamblzweather.model.repositories.about;
+
+import com.ringov.yamblzweather.model.repositories.base.BaseRepository;
+
+/**
+ * Created by ringov on 12.07.17.
+ */
+
+public interface AboutRepository extends BaseRepository {
+}
