+++ app/src/main/java/com/ringov/yamblzweather/model/repositories/about/AboutRepositoryImpl.java
@@ -1,0 +1,13 @@
+package com.ringov.yamblzweather.model.repositories.about;
+
+import com.ringov.yamblzweather.model.repositories.base.BaseRepositoryImpl;
+
+import javax.inject.Singleton;
+
+/**
+ * Created by ringov on 12.07.17.
+ */
+
+@Singleton
+public class AboutRepositoryImpl extends BaseRepositoryImpl implements AboutRepository {
+}
