+++ app/src/main/java/com/ringov/yamblzweather/model/repositories/base/BaseRepository.java
@@ -1,0 +1,8 @@
+package com.ringov.yamblzweather.model.repositories.base;
+
+/**
+ * Created by ringov on 12.07.17.
+ */
+
+public interface BaseRepository {
+}
