+++ app/src/main/java/com/ringov/yamblzweather/model/repositories/base/BaseRepositoryImpl.java
@@ -1,0 +1,8 @@
+package com.ringov.yamblzweather.model.repositories.base;
+
+/**
+ * Created by ringov on 12.07.17.
+ */
+
+public class BaseRepositoryImpl implements BaseRepository {
+}
