+++ app/src/main/java/com/ringov/yamblzweather/model/repositories/weather/WeatherRepository.java
@@ -1,0 +1,14 @@
+package com.ringov.yamblzweather.model.repositories.weather;
+
+import com.ringov.yamblzweather.model.repositories.base.BaseRepository;
+import com.ringov.yamblzweather.presenter.model.WeatherInfo;
+
+import io.reactivex.Observable;
+
+/**
+ * Created by ringov on 12.07.17.
+ */
+
+public interface WeatherRepository extends BaseRepository {
+    Observable<WeatherInfo> getWeatherInfo();
+}
