+++ app/src/main/java/com/ringov/yamblzweather/presenter/about/AboutViewModel.java
@@ -1,0 +1,10 @@
+package com.ringov.yamblzweather.presenter.about;
+
+import com.ringov.yamblzweather.presenter.base.BaseViewModel;
+
+/**
+ * Created by ringov on 12.07.17.
+ */
+
+public class AboutViewModel extends BaseViewModel {
+}
