+++ app/src/main/java/com/ringov/yamblzweather/ui/weather/WeatherFragment.java
@@ -6,8 +6,8 @@ import android.view.View;
 import android.widget.TextView;
 import com.ringov.yamblzweather.R;
 import com.ringov.yamblzweather.ui.base.ModelViewFragment;
-import com.ringov.yamblzweather.viewmodel.model.WeatherInfo;
-import com.ringov.yamblzweather.viewmodel.weather.WeatherViewModel;
+import com.ringov.yamblzweather.presenter.model.WeatherInfo;
+import com.ringov.yamblzweather.presenter.weather.WeatherViewModel;
 
 import butterknife.BindView;
