--- app/src/main/java/com/ringov/yamblzweather/viewmodel/weather/WeatherViewModel.java
@@ -1,25 +1,0 @@
-package com.ringov.yamblzweather.viewmodel.weather;
-
-import com.ringov.yamblzweather.App;
-import com.ringov.yamblzweather.model.weather.WeatherRepository;
-import com.ringov.yamblzweather.viewmodel.base.BaseLiveData;
-import com.ringov.yamblzweather.viewmodel.base.BaseViewModel;
-import com.ringov.yamblzweather.viewmodel.model.WeatherInfo;
-
-import javax.inject.Inject;
-
-/**
- * Created by ringov on 12.07.17.
- */
-
-public class WeatherViewModel extends BaseViewModel<BaseLiveData<WeatherInfo>, WeatherInfo> {
-
-    @Inject
-    WeatherRepository repository;
-
-    public WeatherViewModel() {
-        App.getComponent().inject(this);
-        addDisposable(repository.getWeatherInfo()
-                .subscribe(weather -> getLiveData().updateValue(weather), this::handleError));
-    }
-}
